test: separates demo components into their own files
